% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_c}
\alias{cl_c}
\title{Count control limits for c chart}
\usage{
cl_c(value, size, cbar = NULL, ruleVarySize = 0.1)
}
\arguments{
\item{value}{numeric vector, data to count control limits}

\item{size}{integer vector with same length of value, sample size}

\item{cbar}{if assigned, cbar will not be counted from raw data}

\item{ruleVarySize}{rule of varying sample size, by percentage}
}
\value{
data.frame contains control limits
}
\description{
Count control limits for c chart
}
