% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_np}
\alias{cl_np}
\title{Count control limits for np chart}
\usage{
cl_np(value, size, npbar = NULL, ruleVarySize = 0.1)
}
\arguments{
\item{value}{numeric vector, data to count control limits}

\item{size}{integer vector with same length of value, sample size}

\item{npbar}{if assigned, npbar will not be counted from raw data}

\item{ruleVarySize}{rule of varying sample size, by percentage}
}
\value{
data.frame contains control limits of p charts
}
\description{
Count control limits for np chart
}
