% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_p}
\alias{cl_p}
\title{Count control limits for p chart}
\usage{
cl_p(value, size, pbar = NULL, ruleVarySize = 0.1)
}
\arguments{
\item{value}{numeric vector, data to count control limits}

\item{size}{integer vector with same length of value, sample size}

\item{pbar}{if assigned, pbar will not be counted from raw data}

\item{ruleVarySize}{rule of varying sample size, by percentage}
}
\value{
data.frame contains control limits of p charts
}
\description{
Count control limits for p chart
}
