% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_u}
\alias{cl_u}
\title{Count control limits for u chart}
\usage{
cl_u(value, size, iu, ubar = NULL, ruleVarySize = 0.1)
}
\arguments{
\item{value}{numeric vector, data to count control limits}

\item{size}{integer vector with same length of value, sample size}

\item{iu}{unit of per inspection}

\item{ubar}{if assigned, ubar will not be counted from raw data}

\item{ruleVarySize}{rule of varying sample size, by percentage}
}
\value{
data.frame contains control limits
}
\description{
Count control limits for u chart
}
