% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_X_mR}
\alias{cl_X_mR}
\title{Count control limits of X-mR chart}
\usage{
cl_X_mR(value)
}
\arguments{
\item{value}{numeric vector, data to count control limits}
}
\value{
data.frame contains control limits
}
\description{
Count control limits of X-mR chart
}
