% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_control_limits.R
\name{cl_Xbar_s}
\alias{cl_Xbar_s}
\title{Count control limits for Xbar-mR-R chart}
\usage{
cl_Xbar_s(value, sample_id)
}
\arguments{
\item{value}{numeric vector, data to count control limits}

\item{sample_id}{character vector with same length of value, id of sample}
}
\value{
data.frame contains control limits
}
\description{
Count control limits for Xbar-mR-R chart
}
