% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{cusum_chart}
\alias{cusum_chart}
\title{plot cusum chart}
\usage{
cusum_chart(df, x_var, y_var, info_var = NULL, group_var = NULL,
  info_names = NULL, sizes = NULL, center = NULL, stDev = NULL,
  decision_interval = 5, se_shift = 1, xlab = NULL,
  ylab = "Cumulative Sum", title = "Cusum Chart")
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{info_var}{character vector, variables for hover text}

\item{group_var}{character vector form the smaple id, should be in info_var.
If NULL, each row represent a sample.}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{sizes}{value, sample size of per group, if null, counted by group_by}

\item{center}{value, center of control chart or group statistic, if null, counted by
input data}

\item{stDev}{within-group standard deviation, if null, counted by input data}

\item{decision_interval}{numeric value, specify the number of standard error of the
summary statistics at which the cumulative sum is out of control}

\item{se_shift}{the amount of shift to detect in the process, measured in standard
error of the summary statistics}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{title}
}
\value{
list contains: cusum chart, decision interval, standard error shift,
        points of rule violated low and hight
}
\description{
plot cusum chart
}
