% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Df_cal_xbar}
\alias{Df_cal_xbar}
\title{Calculate Xbar, R, s and mR}
\usage{
Df_cal_xbar(df, x_var, y_var, group_var, info_var, color_var)
}
\arguments{
\item{df}{dataframe for calculating Xbar, R, s and mR}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{group_var}{character vector form the smaple id}

\item{info_var}{character vector, variables for hover text}

\item{color_var}{character, variable of coloring points}
}
\value{
dataframe with summarized xbar, R, s and mR
}
\description{
Calculate Xbar, R, s and mR
}
