% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{ewma_chart}
\alias{ewma_chart}
\title{plot cusum chart}
\usage{
ewma_chart(df, x_var, y_var, info_var = NULL, group_var = NULL,
  info_names = NULL, sizes = NULL, center = NULL, stDev = NULL,
  lambda = 0.2, nsigmas = 3, xlab = NULL, ylab = "EWMA",
  title = "EWMN Chart")
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{info_var}{character vector, variables for hover text}

\item{group_var}{character vector form the smaple id, should be in info_var.
If NULL, each row represent a sample.}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{sizes}{value, sample size of per group, if null, counted by group_by}

\item{center}{value, center of control chart or group statistic, if null, counted by
input data}

\item{stDev}{within-group standard deviation, if null, counted by input data}

\item{lambda}{the smoothing parameter}

\item{nsigmas}{the number of sigmas for computing control limits}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{title}
}
\value{
list contains: cusum chart, lambda, nsigamas, points of rule violated
}
\description{
plot cusum chart
}
