% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Plotly_cusum}
\alias{Plotly_cusum}
\title{plotly cusum chart}
\usage{
Plotly_cusum(x, cusumx, df_info = NULL, info_names = NULL,
  xlab = NULL, ylab = "Cumulative Sum", title = "Cusum Chart")
}
\arguments{
\item{x}{numeric vector of original data points}

\item{cusumx}{cusum object}

\item{df_info}{data frame contians the information to show in plot}

\item{info_names}{character vector, 順序要對應到 df_info 的欄位順序}

\item{xlab}{label for x-axis}

\item{ylab}{label of y-axis}

\item{title}{title of plot}
}
\value{
plotly plot
}
\description{
plotly cusum chart
}
