% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Plotly_ewma}
\alias{Plotly_ewma}
\title{plotly EWMA}
\usage{
Plotly_ewma(x, ewmax, df_info = NULL, info_names = NULL, xlab = NULL,
  ylab = "EWMA", title = "EWMA Chart")
}
\arguments{
\item{x}{numeric vector of original data points}

\item{ewmax}{ewma object}

\item{df_info}{data frame contians the information to show in plot}

\item{info_names}{character vector, 順序要對應到 df_info 的欄位順序}

\item{xlab}{label for x-axis}

\item{ylab}{label of y-axis}

\item{title}{title of plot}
}
\value{
plotly plot
}
\description{
plotly EWMA
}
