% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Plotly_spc}
\alias{Plotly_spc}
\title{Plotly SPC chart}
\usage{
Plotly_spc(x, y, xlab = NULL, ylab = NULL, df_info = NULL,
  info_names = NULL, df_color = NULL, center = NULL, UCL = NULL,
  LCL = NULL, oo_limits = TRUE, violate_runs = TRUE,
  zoneStrip = FALSE, legend = FALSE, color_set = "Set1")
}
\arguments{
\item{x}{vector of x- data}

\item{y}{vector of y- data}

\item{xlab}{character, label of x-axis, showing in hover text}

\item{ylab}{character, label of y-axis, showing in hover text}

\item{df_info}{data frame contians the information to show in plot}

\item{info_names}{character vector, 順序要對應到 df_info 的欄位順序}

\item{df_color}{dataframe or vector, for coloring the points of plot}

\item{center}{center line of control chart}

\item{UCL}{upper control limit}

\item{LCL}{lowe control limit}

\item{oo_limits}{TRUE for count out of limits}

\item{violate_runs}{TRUE for count violating runs}

\item{zoneStrip}{TRUE is want plot strip line for zone A, B, C}

\item{legend}{TRUE is want to show legend on plot}

\item{color_set}{custom color set for plotting}
}
\value{
list contains plotly spc chart, and dataframe with ooc labels
}
\description{
Plotly SPC chart
}
