% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{u_chart}
\alias{u_chart}
\title{u/c chart}
\usage{
u_chart(df, x_var, y_var, size_var, iu = NULL, info_var = NULL,
  color_var = NULL, info_names = NULL, center_line = NULL,
  xlab = NULL, ylab = NULL, title = "", c_chart = FALSE,
  ruleVarySize = 0.1, legend = FALSE, color_set = "Set1")
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable name for x-axis}

\item{y_var}{character, variable name for y-axis, e.g. defect counts}

\item{size_var}{character, variable name for sample size}

\item{iu}{unit per inspection}

\item{info_var}{character vector, variables for hover text}

\item{color_var}{character, variable of coloring points, should be in info_var}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{center_line}{center of control chart}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{title}

\item{c_chart}{logic, TRUE for c-chart}

\item{ruleVarySize}{rule of varying sample size, by percentage}

\item{legend}{logic, TRUE to show legend on plot}

\item{color_set}{custom color set}
}
\value{
SPC chart by plotly
}
\description{
u/c chart
}
