% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{X_mR}
\alias{X_mR}
\title{X-mR chart}
\usage{
X_mR(df, x_var, y_var, info_var = NULL, color_var = NULL,
  info_names = NULL, control_limits = NULL, xlab = NULL,
  ylab_X = "X", ylab_mR = "mR", title = "", mR_chart = TRUE)
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{info_var}{character vector, variables for hover text}

\item{color_var}{character, variable of coloring points, should be in info_var}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{control_limits}{information of control limits of Xbar-R chart: CLx, CLmr, UCLx,
LCLx, UCLmr, LCLmr}

\item{xlab}{x-axis label}

\item{ylab_X}{y-axis label for X chart}

\item{ylab_mR}{y-axis label for mR chart}

\item{title}{title}

\item{mR_chart}{logic, TRUE to include mR chart, FALSE for X chart only}
}
\value{
SPC chart by plotly
}
\description{
X-mR chart
}
