% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Xbar_s}
\alias{Xbar_s}
\title{Xbar-s chart}
\usage{
Xbar_s(df, x_var, y_var, info_var, group_var, color_var = NULL,
  info_names = NULL, control_limits = NULL, xlab = NULL,
  ylab_X = "Xbar", ylab_s = "s", title = "")
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{info_var}{character vector, variables for hover text}

\item{group_var}{character vector form the smaple id, should be in info_var}

\item{color_var}{character, variable of coloring points, should be in info_var}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{control_limits}{information of control limits of Xbar-R chart: CLx, CLs, UCLx,
LCLx, UCLs, LCLs}

\item{xlab}{x-axis label}

\item{ylab_X}{y-axis label for Xbar chart}

\item{ylab_s}{y-axis label for s chart}

\item{title}{title}
}
\value{
SPC chart by plotly
}
\description{
Xbar-s chart
}
