% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_charts.R
\name{Xbar_mR_R}
\alias{Xbar_mR_R}
\title{Xbar-mR-R chart}
\usage{
Xbar_mR_R(df, x_var, y_var, info_var = NULL, group_var = NULL,
  color_var = NULL, info_names = NULL, control_limits = NULL,
  xlab = NULL, ylab_X = "Xbar", ylab_mR = "mR", ylab_R = "R",
  title = "", R_chart = TRUE, legend = TRUE, color_set = "Set1",
  zoneStrip = TRUE)
}
\arguments{
\item{df}{dataframe for plotting, including x, y variables and other variables for
grouping, coloring and hover text.}

\item{x_var}{character, variable names for x-axis}

\item{y_var}{character, variable names for y-axis}

\item{info_var}{character vector, variables for hover text}

\item{group_var}{character vector form the smaple id}

\item{color_var}{character, variable of coloring points}

\item{info_names}{character vector, with the same length of info_var, to replace the
names of info_var for more meaning text for hover text}

\item{control_limits}{information of control limits of Xbar-R chart: CLx, CLmr, CLr,
UCLx, LCLx, UCLr, LCLr, UCLmr, LCLmr}

\item{xlab}{x-axis label}

\item{ylab_X}{y-axis label for Xbar chart}

\item{ylab_mR}{y label of mR chart}

\item{ylab_R}{y-axis label for R chart}

\item{title}{title}

\item{R_chart}{logic, TRUE to include R chart, FALSE for Xbar chart only}

\item{legend}{logic, TRUE to show legend on plot}

\item{color_set}{custom color set}

\item{zoneStrip}{logic, TRUE to show zone strip lines}
}
\value{
SPC chart by plotly
}
\description{
Xbar-mR-R chart
}
